django/requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
djangorestframework==3.13.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
packaging==21.3
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
uritemplate==4.1.1
urllib3==1.26.8
